+++ extensions/yii/twig/CHANGELOG.md
@@ -5,6 +5,7 @@ Yii Framework 2 twig extension Change Log
 ----------------------------
 
 - no changes in this release.
+- Add File based Twig loader for better caching and usability of twig's file based function
 
 2.0.0 alpha, December 1, 2013
 -----------------------------